Move the BatchIf overload tests into BatchIfFixture
They belong with the rest of the operator's coverage rather than in a fixture
of their own, and they can use the source and scheduler already set up there.
